requirements.txt
altair==3.3.0
argh==0.26.2
astor==0.8.0
attrs==19.3.0
backcall==0.1.0
base58==1.0.3
blinker==1.4
boto3==1.10.32
botocore==1.13.32
certifi==2019.9.11
chardet==3.0.4
Click==7.0
colorama==0.4.3
config==0.4.2
cycler==0.10.0
dash==1.9.0
dash-core-components==1.8.0
dash-daq==0.3.1
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.0
decorator==4.4.1
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
fhnw-ds-hs2019-weatherstation-api==0.19
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
helpers==0.2.0
idna==2.8
importlib-metadata==1.1.0
influxdb==5.2.3
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.1
jsonschema==3.2.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
more-itertools==8.0.0
numpy==1.18.0
pandas==1.0.3
parso==0.6.2
pathtools==0.1.2
patsy==0.5.1
pickleshare==0.7.5
Pillow==6.2.1
plotly==4.3.0
plotly-express==0.4.1
prompt-toolkit==3.0.5
protobuf==3.11.1
psycopg2==2.8.4
Pygments==2.6.1
pyparsing==2.4.5
pyrsistent==0.15.6
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.2
requests==2.22.0
retrying==1.3.3
s3transfer==0.2.1
scikit-learn==0.22.2.post1
scipy==1.3.3
seaborn==0.9.0
six==1.12.0
statsmodels==0.10.2
streamlit==0.51.0
toml==0.10.0
toolz==0.10.0
tornado==5.1.1
tqdm==4.46.0
traitlets==4.3.3
tzlocal==2.0.0
urllib3==1.25.6
validators==0.14.1
watchdog==0.9.0
wcwidth==0.1.9
Werkzeug==0.16.0
zipp==0.6.0
